Add and reorg. declarative unit tests.
