+++ test/declarative/__init__.py
@@ -5,3 +5,5 @@ from .test_declarative_dict import *
 from .test_declarative_regex import *
 from .test_declarative_when import *
 from .test_declarative_dict_parse import *
+from .test_declarative_and import *
+from .test_declarative_add import *
