+++ test/declarative/test_add_02.yaml
@@ -1,0 +1,5 @@
+---
+!ADD
+- 1
+- 2
+
