+++ test/declarative/test_declarative_join.py
@@ -41,14 +41,3 @@ class TestDeclarativeJoin(bspump.unittest.TestCase):
 		res = decl({}, event)
 		self.assertEqual(res, None)
 
-
-	def test_add_01(self):
-		event = {
-			'string1': "STRING1",
-			'string2': "STRING2",
-		}
-
-		decl = self.load('./test_add_01.yaml')
-
-		res = decl({}, event)
-		self.assertEqual(res, "STRING1STRING2")
